+++ src/process_pipeline.h
@@ -8,7 +8,7 @@ typedef struct processPipeline ProcessPipeline;
 
 // Makes a pipeline that will read from input_fd and write to output_fd, passing the data
 // through added processes
-ProcessPipeline *pp_new(size_t process_num, int input_fd, int output_fd);
+ProcessPipeline *pp_new(int process_num, int input_fd, int output_fd);
 
 void pp_free(ProcessPipeline *pp);
 
@@ -22,7 +22,7 @@ bool pp_wait_check_all(ProcessPipeline *pp);
 // Returns number of finished processes or -1 if child returned non zero
 int pp_check_end_num(ProcessPipeline *pp);
 
-size_t pp_get_process_num(ProcessPipeline *pp);
+int pp_get_process_num(ProcessPipeline *pp);
 
 
 #endif
